redfish: control: read power-cap properties as uint32_t
getChassisPower() read the xyz.openbmc_project.Control.Power.Cap
properties (MaxPowerCapValue, MinPowerCapValue, PowerCap/SetPoint,
DefaultPowerCap) with std::get_if<size_t>.

DbusVariantType has no size_t alternative, so std::get_if<size_t> only
compiles because size_t aliases onto uint32_t on 32-bit targets and
uint64_t on 64-bit ones. The read therefore resolves to a different
D-Bus type depending on the architecture.

Every property name this code reads is uint32. MaxPowerCapValue,
MinPowerCapValue, PowerCap and DefaultPowerCap are declared uint32 by
the Control.Power.Cap PDI. PowerCapPercentage, the setpoint when
power-control-type-percentage is enabled, is not a PDI property but
is published as uint32 by nvidia-power-manager.

On 32-bit the alias happened to match, so the handler worked. On
64-bit std::get_if returns nullptr and

    GET /redfish/v1/Chassis/<id>/Controls/<control>

fails with Base.1.19.InternalError (500).

Read the properties as uint32_t instead, matching the way
utils/environment_util.hpp already reads the same properties. On
32-bit this is a no-op; on 64-bit it fixes the 500.

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-24925

+++ redfish-core/lib/control.hpp
@@ -141,8 +141,9 @@ inline void getChassisPower(const std::shared_ptr<bmcweb::AsyncResp>& asyncResp,
                                     if (propertyName == "MaxPowerCapValue")
                                     {
                                         propertyName = "AllowableMax";
-                                        const auto* value = std::get_if<size_t>(
-                                            &property.second);
+                                        const auto* value =
+                                            std::get_if<uint32_t>(
+                                                &property.second);
                                         if (value == nullptr)
                                         {
                                             BMCWEB_LOG_ERROR(
@@ -158,8 +159,9 @@ inline void getChassisPower(const std::shared_ptr<bmcweb::AsyncResp>& asyncResp,
                                     if (propertyName == "MinPowerCapValue")
                                     {
                                         propertyName = "AllowableMin";
-                                        const auto* value = std::get_if<size_t>(
-                                            &property.second);
+                                        const auto* value =
+                                            std::get_if<uint32_t>(
+                                                &property.second);
                                         if (value == nullptr)
                                         {
                                             BMCWEB_LOG_ERROR(
@@ -175,8 +177,9 @@ inline void getChassisPower(const std::shared_ptr<bmcweb::AsyncResp>& asyncResp,
                                     if (propertyName == setPointPropName())
                                     {
                                         propertyName = "SetPoint";
-                                        const auto* value = std::get_if<size_t>(
-                                            &property.second);
+                                        const auto* value =
+                                            std::get_if<uint32_t>(
+                                                &property.second);
                                         if (value == nullptr)
                                         {
                                             BMCWEB_LOG_ERROR(
@@ -192,8 +195,9 @@ inline void getChassisPower(const std::shared_ptr<bmcweb::AsyncResp>& asyncResp,
                                     if (propertyName == "DefaultPowerCap")
                                     {
                                         propertyName = "DefaultSetPoint";
-                                        const auto* value = std::get_if<size_t>(
-                                            &property.second);
+                                        const auto* value =
+                                            std::get_if<uint32_t>(
+                                                &property.second);
                                         if (value == nullptr)
                                         {
                                             BMCWEB_LOG_ERROR(
